migrate UT 'address_book'
