+++ unit-tests/address_book/CMakeLists.txt
@@ -39,48 +39,81 @@ ledger_unit_tests_add_test(
     COMPILE_DEFS ${AB_COMPILE_DEFS}
 )
 
+# provide_contact split: one binary per compilation unit under test
 ledger_unit_tests_add_test(
-    NAME         test_address_book_provide_contact
-    SOURCES      ${AB_SOURCES_COMMON}
-                 tlv_test_helpers.c
-                 ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
-                 ${SDK_DIR}/app_features/address_book/src/identity_provide_contact.c
-                 ${SDK_DIR}/app_features/address_book/src/ledger_account_provide_contact.c
-    MOCK_HEADERS ${SDK_DIR}/lib_standard_app/io.h
-                 ${SDK_DIR}/include/ledger_assert_internals.h
-                 ${SDK_DIR}/include/os_address_book.h
-                 ${SDK_DIR}/include/os_utils.h
-                 ${SDK_DIR}/include/os_pic.h
-                 ${SDK_DIR}/lib_nbgl/include/nbgl_use_case.h
-                 ${SDK_DIR}/lib_cxng/include/lcx_rng.h
-    INCLUDE_DIRS ${AB_INCLUDE_DIRS}
-    COMPILE_DEFS ${AB_COMPILE_DEFS}
+    NAME            test_identity_provide_contact
+    SOURCES         ${AB_SOURCES_COMMON}
+                    tlv_test_helpers.c
+                    ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
+                    ${SDK_DIR}/app_features/address_book/src/identity_provide_contact.c
+    MOCK_HEADERS    ${SDK_DIR}/lib_standard_app/io.h
+                    ${SDK_DIR}/include/ledger_assert_internals.h
+                    ${SDK_DIR}/include/os_address_book.h
+                    ${SDK_DIR}/include/os_utils.h
+                    ${SDK_DIR}/include/os_pic.h
+                    ${SDK_DIR}/lib_nbgl/include/nbgl_use_case.h
+                    ${SDK_DIR}/lib_cxng/include/lcx_rng.h
+    INCLUDE_DIRS    ${AB_INCLUDE_DIRS}
+    COMPILE_DEFS    ${AB_COMPILE_DEFS}
+    COMPILE_OPTIONS -Wno-array-parameter
 )
-# CMock generates pointer params from array-typed declarations in os_address_book.h;
-# the mismatch is benign but triggers -Warray-parameter in GCC 11+.
-target_compile_options(test_address_book_provide_contact PRIVATE -Wno-array-parameter)
 
 ledger_unit_tests_add_test(
-    NAME         test_address_book_ui_flows
-    SOURCES      ${AB_SOURCES_COMMON}
-                 tlv_test_helpers.c
-                 ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
-                 ${SDK_DIR}/app_features/address_book/src/identity_register.c
-                 ${SDK_DIR}/app_features/address_book/src/identity_edit_contact_name.c
-                 ${SDK_DIR}/app_features/address_book/src/identity_edit_identifier.c
-                 ${SDK_DIR}/app_features/address_book/src/identity_edit_scope.c
-                 ${SDK_DIR}/app_features/address_book/src/ledger_account_register.c
-                 ${SDK_DIR}/app_features/address_book/src/ledger_account_edit.c
-    MOCK_HEADERS ${SDK_DIR}/lib_standard_app/io.h
-                 ${SDK_DIR}/include/ledger_assert_internals.h
-                 ${SDK_DIR}/include/os_address_book.h
-                 ${SDK_DIR}/include/os_utils.h
-                 ${SDK_DIR}/include/os_pic.h
-                 ${SDK_DIR}/lib_cxng/include/lcx_rng.h
-    INCLUDE_DIRS ${AB_INCLUDE_DIRS}
-    COMPILE_DEFS ${AB_COMPILE_DEFS}
+    NAME            test_ledger_account_provide_contact
+    SOURCES         ${AB_SOURCES_COMMON}
+                    tlv_test_helpers.c
+                    ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
+                    ${SDK_DIR}/app_features/address_book/src/ledger_account_provide_contact.c
+    MOCK_HEADERS    ${SDK_DIR}/lib_standard_app/io.h
+                    ${SDK_DIR}/include/ledger_assert_internals.h
+                    ${SDK_DIR}/include/os_address_book.h
+                    ${SDK_DIR}/include/os_utils.h
+                    ${SDK_DIR}/include/os_pic.h
+                    ${SDK_DIR}/lib_nbgl/include/nbgl_use_case.h
+                    ${SDK_DIR}/lib_cxng/include/lcx_rng.h
+    INCLUDE_DIRS    ${AB_INCLUDE_DIRS}
+    COMPILE_DEFS    ${AB_COMPILE_DEFS}
+    COMPILE_OPTIONS -Wno-array-parameter
 )
-target_compile_options(test_address_book_ui_flows PRIVATE -Wno-array-parameter)
+
+# ui_flows split: identity flows vs ledger-account flows
+ledger_unit_tests_add_test(
+    NAME            test_identity_ui_flows
+    SOURCES         ${AB_SOURCES_COMMON}
+                    tlv_test_helpers.c
+                    ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
+                    ${SDK_DIR}/app_features/address_book/src/identity_register.c
+                    ${SDK_DIR}/app_features/address_book/src/identity_edit_contact_name.c
+                    ${SDK_DIR}/app_features/address_book/src/identity_edit_identifier.c
+                    ${SDK_DIR}/app_features/address_book/src/identity_edit_scope.c
+    MOCK_HEADERS    ${SDK_DIR}/lib_standard_app/io.h
+                    ${SDK_DIR}/include/ledger_assert_internals.h
+                    ${SDK_DIR}/include/os_address_book.h
+                    ${SDK_DIR}/include/os_utils.h
+                    ${SDK_DIR}/include/os_pic.h
+                    ${SDK_DIR}/lib_cxng/include/lcx_rng.h
+    INCLUDE_DIRS    ${AB_INCLUDE_DIRS}
+    COMPILE_DEFS    ${AB_COMPILE_DEFS}
+    COMPILE_OPTIONS -Wno-array-parameter
+)
+
+ledger_unit_tests_add_test(
+    NAME            test_ledger_account_ui_flows
+    SOURCES         ${AB_SOURCES_COMMON}
+                    tlv_test_helpers.c
+                    ${SDK_DIR}/app_features/address_book/src/address_book_crypto.c
+                    ${SDK_DIR}/app_features/address_book/src/ledger_account_register.c
+                    ${SDK_DIR}/app_features/address_book/src/ledger_account_edit.c
+    MOCK_HEADERS    ${SDK_DIR}/lib_standard_app/io.h
+                    ${SDK_DIR}/include/ledger_assert_internals.h
+                    ${SDK_DIR}/include/os_address_book.h
+                    ${SDK_DIR}/include/os_utils.h
+                    ${SDK_DIR}/include/os_pic.h
+                    ${SDK_DIR}/lib_cxng/include/lcx_rng.h
+    INCLUDE_DIRS    ${AB_INCLUDE_DIRS}
+    COMPILE_DEFS    ${AB_COMPILE_DEFS}
+    COMPILE_OPTIONS -Wno-array-parameter
+)
 
 ledger_unit_tests_add_test(
     NAME         test_address_book_apdu
