+++ unit-tests/address_book/test_identity_ui_flows.c
@@ -6,31 +6,17 @@
  *****************************************************************************/
 
 /**
- * @file test_address_book_ui_flows.c
- * @brief Unit tests for the six address-book commands that launch a NBGL
- *        confirmation UI before sending a response.
- *
- * nbgl_useCaseReviewLight() and display_register_ledger_account_review() are
- * stubbed to invoke their callback synchronously with g_mock_review_choice
- * (true = confirm, false = reject), making the full confirm/reject paths
- * reachable in unit tests.  All error paths (bad TLV, missing field, HMAC
- * failure, app-callback rejection) return early before reaching the UI.
+ * @file test_identity_ui_flows.c
+ * @brief Unit tests for the four identity address-book commands that launch NBGL UI.
  *
  * Functions under test:
- *   register_identity()         identity_register.c          TYPE 0x2d
- *   edit_contact_name()         identity_edit_contact_name.c TYPE 0x2e
- *   edit_identifier()           identity_edit_identifier.c   TYPE 0x31
- *   edit_scope()                identity_edit_scope.c        TYPE 0x32
- *   register_ledger_account()   ledger_account_register.c    TYPE 0x2f
- *   edit_ledger_account()       ledger_account_edit.c        TYPE 0x30
+ *   register_identity()    identity_register.c          TYPE 0x2d
+ *   edit_contact_name()    identity_edit_contact_name.c TYPE 0x2e
+ *   edit_identifier()      identity_edit_identifier.c   TYPE 0x31
+ *   edit_scope()           identity_edit_scope.c        TYPE 0x32
  *
- * Stubs provided here:
- *   handle_check_register_identity()       — configurable via g_mock_ri_result
- *   handle_check_edit_identifier()         — configurable via g_mock_ei_result
- *   handle_check_register_ledger_account() — configurable via g_mock_rla_result
- *   handle_check_edit_ledger_account()     — configurable via g_mock_ela_result
- *   display_register_ledger_account_review() — calls callback(g_mock_review_choice)
- *   LARGE_ADDRESS_BOOK_ICON                — zero-initialised icon
+ * nbgl_useCaseReviewLight() is stubbed to invoke its callback synchronously
+ * with g_mock_review_choice, making confirm/reject paths reachable in tests.
  */
 
 #include <stdbool.h>
@@ -44,29 +30,20 @@
 #include "Mockos_utils.h"
 #include "Mocklcx_rng.h"
 
-#include "nbgl_types.h" /* nbgl_icon_details_t */
+#include "nbgl_types.h"
 
-/* Types from nbgl_use_case.h / nbgl_content.h — forward-declared to avoid
- * pulling in the full nbgl_use_case.h header. */
 typedef uint32_t nbgl_operationType_t;
 typedef void (*nbgl_callback_t)(void);
 typedef void (*nbgl_choiceCallback_t)(bool confirm);
 typedef struct nbgl_contentTagValue_s     nbgl_contentTagValue_t;
 typedef struct nbgl_contentTagValueList_s nbgl_contentTagValueList_t;
 
-/* IO buffers referenced by address_book_crypto.c. */
 unsigned char G_io_rx_buffer[OS_IO_SEPH_BUFFER_SIZE + 1];
 unsigned char G_io_tx_buffer[OS_IO_SEPH_BUFFER_SIZE + 1];
 
-/* Icon stub: LARGE_ADDRESS_BOOK_ICON expands to C_Address_Book_64px on Flex/Stax. */
 const nbgl_icon_details_t C_Address_Book_64px = {0};
 
-/* Controlled by tests: true = user confirms, false = user rejects. */
 bool g_mock_review_choice = true;
-
-/* NBGL stubs — nbgl_use_case.h is NOT in MOCK_HEADERS so we provide custom
- * implementations that invoke the callback synchronously, allowing the full
- * confirm/reject code paths to be exercised in unit tests. */
 
 void nbgl_useCaseReviewLight(nbgl_operationType_t              operationType,
                              const nbgl_contentTagValueList_t *tagValueList,
@@ -93,16 +70,13 @@ void nbgl_useCaseStatus(const char *message, bool isSuccess, nbgl_callback_t qui
 }
 
 #include "identity.h"
-#include "ledger_account.h"
 #include "status_words.h"
 #include "tlv_test_helpers.h"
 
 /* ── Mock controls ───────────────────────────────────────────────────────── */
-/* Local app-entrypoint stubs */
-static bool g_mock_ri_result  = true; /* handle_check_register_identity      */
-static bool g_mock_ei_result  = true; /* handle_check_edit_identifier         */
-static bool g_mock_rla_result = true; /* handle_check_register_ledger_account */
-static bool g_mock_ela_result = true; /* handle_check_edit_ledger_account     */
+
+static bool g_mock_ri_result = true;
+static bool g_mock_ei_result = true;
 
 void setUp(void)
 {
@@ -118,8 +92,6 @@ void setUp(void)
     g_mock_review_choice = true;
     g_mock_ri_result     = true;
     g_mock_ei_result     = true;
-    g_mock_rla_result    = true;
-    g_mock_ela_result    = true;
 }
 
 void tearDown(void)
@@ -148,30 +120,10 @@ bool handle_check_edit_identifier(const edit_identifier_t *params)
     return g_mock_ei_result;
 }
 
-bool handle_check_register_ledger_account(ledger_account_t *params)
-{
-    (void) params;
-    return g_mock_rla_result;
-}
-
-bool handle_check_edit_ledger_account(edit_ledger_account_t *params)
-{
-    (void) params;
-    return g_mock_ela_result;
-}
-
-/* Coin-app review entrypoint for Ledger Account flows. */
-void display_register_ledger_account_review(nbgl_choiceCallback_t callback)
-{
-    callback(g_mock_review_choice);
-}
-
-/* ── Finalise / apply callbacks (called from review_choice) ──────────────── */
 void finalize_ui_register_identity(void) {}
 void finalize_ui_edit_contact_name(void) {}
 void finalize_ui_edit_identifier(void) {}
 void finalize_ui_edit_scope(void) {}
-void finalize_ui_ledger_account(void) {}
 
 void on_edit_contact_name_applied(const edit_contact_name_t *e)
 {
@@ -185,17 +137,13 @@ void on_edit_scope_applied(const edit_scope_t *e)
 {
     (void) e;
 }
-void on_edit_ledger_account_applied(const edit_ledger_account_t *e)
-{
-    (void) e;
-}
 
-/* Dynamic tag-value callbacks for paginated NBGL review — unused in unit tests. */
 nbgl_contentTagValue_t *get_register_identity_tagValue(uint8_t idx)
 {
     (void) idx;
     return NULL;
 }
+
 nbgl_contentTagValue_t *get_edit_identifier_tagValue(uint8_t idx)
 {
     (void) idx;
@@ -204,10 +152,6 @@ nbgl_contentTagValue_t *get_edit_identifier_tagValue(uint8_t idx)
 
 /* ══════════════════════════════════════════════════════════════════════════
  * 1. register_identity  (TYPE_REGISTER_IDENTITY = 0x2d)
- *
- * Tags: 0x01 type, 0x02 version, 0xf0 name, 0xf1 scope, 0xf2 identifier,
- *       0x69 deriv_path, 0x51 family, [0x23 chain_id],
- *       [0xf6 group_handle + 0x29 hmac_proof]  ← optional pair
  * ══════════════════════════════════════════════════════════════════════════ */
 
 static size_t build_register_identity(uint8_t    *buf,
@@ -256,23 +200,10 @@ static size_t build_register_identity(uint8_t    *buf,
     return off;
 }
 
-/* Valid BTC registration without the optional group_handle/hmac_proof pair */
 static size_t build_valid_register_identity(uint8_t *buf, size_t sz)
 {
-    return build_register_identity(buf,
-                                   sz,
-                                   0x2d,
-                                   0x01,
-                                   "Alice",
-                                   "Bitcoin",
-                                   true,  /* identifier   */
-                                   true,  /* deriv_path   */
-                                   true,  /* family       */
-                                   0x00,  /* FAMILY_BITCOIN */
-                                   false, /* no chain_id  */
-                                   false, /* no group_handle */
-                                   false  /* no hmac_proof */
-    );
+    return build_register_identity(
+        buf, sz, 0x2d, 0x01, "Alice", "Bitcoin", true, true, true, 0x00, false, false, false);
 }
 
 static void test_ri_wrong_struct_type(void)
@@ -286,49 +217,23 @@ static void test_ri_wrong_struct_type(void)
 static void test_ri_missing_mandatory_field(void)
 {
     uint8_t buf[512];
-    /* omit scope → verify_fields fails */
-    size_t len = build_register_identity(buf,
-                                         sizeof(buf),
-                                         0x2d,
-                                         0x01,
-                                         "Alice",
-                                         NULL, /* no scope */
-                                         true,
-                                         true,
-                                         true,
-                                         0x00,
-                                         false,
-                                         false,
-                                         false);
+    size_t  len = build_register_identity(
+        buf, sizeof(buf), 0x2d, 0x01, "Alice", NULL, true, true, true, 0x00, false, false, false);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_identity(buf, len));
 }
 
 static void test_ri_group_handle_without_hmac_proof(void)
 {
     uint8_t buf[512];
-    /* group_handle present but hmac_proof absent → verify_fields rejects pair */
-    size_t len = build_register_identity(buf,
-                                         sizeof(buf),
-                                         0x2d,
-                                         0x01,
-                                         "Alice",
-                                         "BTC",
-                                         true,
-                                         true,
-                                         true,
-                                         0x00,
-                                         false,
-                                         true, /* group_handle */
-                                         false /* no hmac_proof */
-    );
+    size_t  len = build_register_identity(
+        buf, sizeof(buf), 0x2d, 0x01, "Alice", "BTC", true, true, true, 0x00, false, true, false);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_identity(buf, len));
 }
 
 static void test_ri_ethereum_missing_chain_id(void)
 {
     uint8_t buf[512];
-    /* FAMILY_ETHEREUM but chain_id omitted → verify_fields rejects */
-    size_t len = build_register_identity(buf,
+    size_t  len = build_register_identity(buf,
                                          sizeof(buf),
                                          0x2d,
                                          0x01,
@@ -337,8 +242,8 @@ static void test_ri_ethereum_missing_chain_id(void)
                                          true,
                                          true,
                                          true,
-                                         0x01,  /* FAMILY_ETHEREUM */
-                                         false, /* no chain_id */
+                                         0x01,
+                                         false,
                                          false,
                                          false);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_identity(buf, len));
@@ -348,8 +253,7 @@ static void test_ri_hmac_fails(void)
 {
     sys_address_book_hmac_verify_IgnoreAndReturn(false);
     uint8_t buf[512];
-    /* include group_handle + hmac_proof so the HMAC check runs */
-    size_t len = build_register_identity(
+    size_t  len = build_register_identity(
         buf, sizeof(buf), 0x2d, 0x01, "Alice", "BTC", true, true, true, 0x00, false, true, true);
     TEST_ASSERT_EQUAL_INT(SWO_SECURITY_CONDITION_NOT_SATISFIED, register_identity(buf, len));
 }
@@ -379,10 +283,6 @@ static void test_ri_review_rejected(void)
 
 /* ══════════════════════════════════════════════════════════════════════════
  * 2. edit_contact_name  (TYPE_EDIT_CONTACT_NAME = 0x2e)
- *
- * Tags: 0x01, 0x02, 0xf0 new_name, 0xf3 prev_name,
- *       0xf6 group_handle, 0x69 deriv_path, 0x29 hmac_proof
- * All mandatory; no family/chain_id; no app callback.
  * ══════════════════════════════════════════════════════════════════════════ */
 
 static size_t build_edit_contact_name(uint8_t    *buf,
@@ -434,16 +334,8 @@ static void test_ecn_wrong_struct_type(void)
 static void test_ecn_missing_mandatory_field(void)
 {
     uint8_t buf[512];
-    /* omit prev_name → verify_fields fails */
-    size_t len = build_edit_contact_name(buf,
-                                         sizeof(buf),
-                                         0x2e,
-                                         0x01,
-                                         "Bob",
-                                         NULL, /* no prev_name */
-                                         true,
-                                         true,
-                                         true);
+    size_t  len
+        = build_edit_contact_name(buf, sizeof(buf), 0x2e, 0x01, "Bob", NULL, true, true, true);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_contact_name(buf, len));
 }
 
@@ -472,12 +364,6 @@ static void test_ecn_review_rejected(void)
 
 /* ══════════════════════════════════════════════════════════════════════════
  * 3. edit_identifier  (TYPE_EDIT_IDENTIFIER = 0x31)
- *
- * Tags: 0x01, 0x02, 0xf0 name, 0xf1 scope, 0xf2 new_identifier,
- *       0xf4 prev_identifier, 0xf6 group_handle, 0x69 deriv_path,
- *       0x51 family, [0x23 chain_id], 0x29 hmac_proof, 0xf7 hmac_rest
- * All non-bracketed mandatory; chain_id only for Ethereum.
- * App callback: handle_check_edit_identifier → SWO_WRONG_PARAMETER_VALUE.
  * ══════════════════════════════════════════════════════════════════════════ */
 
 static size_t build_edit_identifier(uint8_t    *buf,
@@ -547,7 +433,7 @@ static size_t build_valid_edit_identifier(uint8_t *buf, size_t sz)
                                  true,
                                  true,
                                  true,
-                                 0x00, /* FAMILY_BITCOIN, no chain_id */
+                                 0x00,
                                  false,
                                  true,
                                  true);
@@ -577,8 +463,7 @@ static void test_ei_wrong_struct_type(void)
 static void test_ei_missing_mandatory_field(void)
 {
     uint8_t buf[512];
-    /* omit hmac_rest → verify_fields fails */
-    size_t len = build_edit_identifier(buf,
+    size_t  len = build_edit_identifier(buf,
                                        sizeof(buf),
                                        0x31,
                                        0x01,
@@ -592,15 +477,14 @@ static void test_ei_missing_mandatory_field(void)
                                        0x00,
                                        false,
                                        true,
-                                       false /* no hmac_rest */);
+                                       false);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_identifier(buf, len));
 }
 
 static void test_ei_ethereum_missing_chain_id(void)
 {
     uint8_t buf[512];
-    /* FAMILY_ETHEREUM but chain_id omitted → verify_fields rejects */
-    size_t len = build_edit_identifier(buf,
+    size_t  len = build_edit_identifier(buf,
                                        sizeof(buf),
                                        0x31,
                                        0x01,
@@ -611,8 +495,8 @@ static void test_ei_ethereum_missing_chain_id(void)
                                        true,
                                        true,
                                        true,
-                                       0x01,  /* FAMILY_ETHEREUM */
-                                       false, /* no chain_id */
+                                       0x01,
+                                       false,
                                        true,
                                        true);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_identifier(buf, len));
@@ -651,11 +535,6 @@ static void test_ei_review_rejected(void)
 
 /* ══════════════════════════════════════════════════════════════════════════
  * 4. edit_scope  (TYPE_EDIT_SCOPE = 0x32)
- *
- * Tags: 0x01, 0x02, 0xf0 name, 0xf1 new_scope, 0xf2 identifier,
- *       0xf5 prev_scope, 0xf6 group_handle, 0x69 deriv_path,
- *       0x51 family, [0x23 chain_id], 0x29 hmac_proof, 0xf7 hmac_rest
- * No app callback — goes straight to UI after HMAC verification.
  * ══════════════════════════════════════════════════════════════════════════ */
 
 static size_t build_edit_scope(uint8_t    *buf,
@@ -725,7 +604,7 @@ static size_t build_valid_edit_scope(uint8_t *buf, size_t sz)
                             true,
                             true,
                             true,
-                            0x00, /* FAMILY_BITCOIN */
+                            0x00,
                             false,
                             true,
                             true);
@@ -755,8 +634,7 @@ static void test_es_wrong_struct_type(void)
 static void test_es_missing_mandatory_field(void)
 {
     uint8_t buf[512];
-    /* omit hmac_rest → verify_fields fails */
-    size_t len = build_edit_scope(buf,
+    size_t  len = build_edit_scope(buf,
                                   sizeof(buf),
                                   0x32,
                                   0x01,
@@ -770,15 +648,14 @@ static void test_es_missing_mandatory_field(void)
                                   0x00,
                                   false,
                                   true,
-                                  false /* no hmac_rest */);
+                                  false);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_scope(buf, len));
 }
 
 static void test_es_ethereum_missing_chain_id(void)
 {
     uint8_t buf[512];
-    /* FAMILY_ETHEREUM but chain_id omitted → verify_fields rejects */
-    size_t len = build_edit_scope(buf,
+    size_t  len = build_edit_scope(buf,
                                   sizeof(buf),
                                   0x32,
                                   0x01,
@@ -789,8 +666,8 @@ static void test_es_ethereum_missing_chain_id(void)
                                   true,
                                   true,
                                   true,
-                                  0x01,  /* FAMILY_ETHEREUM */
-                                  false, /* no chain_id */
+                                  0x01,
+                                  false,
                                   true,
                                   true);
     TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_scope(buf, len));
@@ -817,233 +694,6 @@ static void test_es_review_rejected(void)
     uint8_t buf[512];
     size_t  len = build_valid_edit_scope(buf, sizeof(buf));
     TEST_ASSERT_EQUAL_INT(SWO_NO_RESPONSE, edit_scope(buf, len));
-}
-
-/* ══════════════════════════════════════════════════════════════════════════
- * 5. register_ledger_account  (TYPE_REGISTER_LEDGER_ACCOUNT = 0x2f)
- *
- * Tags: 0x01, 0x02, 0xf0 name, 0x69 deriv_path, 0x51 family,
- *       [0x23 chain_id]
- * No HMAC; app callback: handle_check_register_ledger_account → SWO_WRONG_PARAMETER_VALUE.
- * On success: display_register_ledger_account_review() then SWO_NO_RESPONSE.
- * ══════════════════════════════════════════════════════════════════════════ */
-
-static size_t build_register_ledger_account(uint8_t    *buf,
-                                            size_t      buf_size,
-                                            uint8_t     type,
-                                            uint8_t     version,
-                                            const char *name,
-                                            bool        include_deriv,
-                                            bool        include_family,
-                                            uint8_t     family,
-                                            bool        include_chain_id)
-{
-    size_t off = 0;
-    (void) buf_size;
-
-    tlv_u8(buf, &off, 0x01, type);
-    tlv_u8(buf, &off, 0x02, version);
-    if (name) {
-        tlv_append(buf, &off, 0xf0, (const uint8_t *) name, (uint8_t) strlen(name));
-    }
-    if (include_deriv) {
-        tlv_append(buf, &off, 0x69, BIP32_ETH_PATH, sizeof(BIP32_ETH_PATH));
-    }
-    if (include_family) {
-        tlv_u8(buf, &off, 0x51, family);
-    }
-    if (include_chain_id) {
-        tlv_append(buf, &off, 0x23, ETH_CHAIN_ID_1, sizeof(ETH_CHAIN_ID_1));
-    }
-    return off;
-}
-
-static size_t build_valid_register_ledger_account(uint8_t *buf, size_t sz)
-{
-    return build_register_ledger_account(buf, sz, 0x2f, 0x01, "MyLedger", true, true, 0x00, false);
-}
-
-static void test_rla_wrong_struct_type(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_register_ledger_account(
-        buf, sizeof(buf), 0xFF, 0x01, "MyLedger", true, true, 0x00, false);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_ledger_account(buf, len));
-}
-
-static void test_rla_missing_mandatory_field(void)
-{
-    uint8_t buf[512];
-    /* omit derivation_path → verify_fields fails */
-    size_t len = build_register_ledger_account(buf,
-                                               sizeof(buf),
-                                               0x2f,
-                                               0x01,
-                                               "MyLedger",
-                                               false, /* no deriv */
-                                               true,
-                                               0x00,
-                                               false);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_ledger_account(buf, len));
-}
-
-static void test_rla_ethereum_missing_chain_id(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_register_ledger_account(buf,
-                                               sizeof(buf),
-                                               0x2f,
-                                               0x01,
-                                               "EthLedger",
-                                               true,
-                                               true,
-                                               0x01, /* FAMILY_ETHEREUM */
-                                               false /* no chain_id */);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, register_ledger_account(buf, len));
-}
-
-static void test_rla_app_rejects(void)
-{
-    g_mock_rla_result = false;
-    uint8_t buf[512];
-    size_t  len = build_valid_register_ledger_account(buf, sizeof(buf));
-    TEST_ASSERT_EQUAL_INT(SWO_WRONG_PARAMETER_VALUE, register_ledger_account(buf, len));
-}
-
-static void test_rla_success(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_valid_register_ledger_account(buf, sizeof(buf));
-    TEST_ASSERT_EQUAL_INT(SWO_NO_RESPONSE, register_ledger_account(buf, len));
-}
-
-static void test_rla_review_rejected(void)
-{
-    g_mock_review_choice = false;
-    uint8_t buf[512];
-    size_t  len = build_valid_register_ledger_account(buf, sizeof(buf));
-    TEST_ASSERT_EQUAL_INT(SWO_NO_RESPONSE, register_ledger_account(buf, len));
-}
-
-/* ══════════════════════════════════════════════════════════════════════════
- * 6. edit_ledger_account  (TYPE_EDIT_LEDGER_ACCOUNT = 0x30)
- *
- * Tags: 0x01, 0x02, 0xf0 new_name, 0xf3 prev_name,
- *       0x69 deriv_path, 0x51 family, [0x23 chain_id], 0x29 hmac_proof
- * HMAC: address_book_verify_hmac_proof_ledger_account (mocked via sys_address_book_hmac_verify).
- * App callback: handle_check_edit_ledger_account → SWO_INCORRECT_DATA on rejection.
- * ══════════════════════════════════════════════════════════════════════════ */
-
-static size_t build_edit_ledger_account(uint8_t    *buf,
-                                        size_t      buf_size,
-                                        uint8_t     type,
-                                        uint8_t     version,
-                                        const char *new_name,
-                                        const char *prev_name,
-                                        bool        include_deriv,
-                                        bool        include_family,
-                                        uint8_t     family,
-                                        bool        include_chain_id,
-                                        bool        include_hmac_proof)
-{
-    size_t off = 0;
-    (void) buf_size;
-
-    tlv_u8(buf, &off, 0x01, type);
-    tlv_u8(buf, &off, 0x02, version);
-    if (new_name) {
-        tlv_append(buf, &off, 0xf0, (const uint8_t *) new_name, (uint8_t) strlen(new_name));
-    }
-    if (prev_name) {
-        tlv_append(buf, &off, 0xf3, (const uint8_t *) prev_name, (uint8_t) strlen(prev_name));
-    }
-    if (include_deriv) {
-        tlv_append(buf, &off, 0x69, BIP32_ETH_PATH, sizeof(BIP32_ETH_PATH));
-    }
-    if (include_family) {
-        tlv_u8(buf, &off, 0x51, family);
-    }
-    if (include_chain_id) {
-        tlv_append(buf, &off, 0x23, ETH_CHAIN_ID_1, sizeof(ETH_CHAIN_ID_1));
-    }
-    if (include_hmac_proof) {
-        tlv_append(buf, &off, 0x29, ZERO_32, sizeof(ZERO_32));
-    }
-    return off;
-}
-
-static size_t build_valid_edit_ledger_account(uint8_t *buf, size_t sz)
-{
-    return build_edit_ledger_account(
-        buf, sz, 0x30, 0x01, "NewLedger", "OldLedger", true, true, 0x00, false, true);
-}
-
-static void test_ela_wrong_struct_type(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_edit_ledger_account(
-        buf, sizeof(buf), 0xFF, 0x01, "New", "Old", true, true, 0x00, false, true);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_ledger_account(buf, len));
-}
-
-static void test_ela_missing_mandatory_field(void)
-{
-    uint8_t buf[512];
-    /* omit hmac_proof → verify_fields fails */
-    size_t len = build_edit_ledger_account(buf,
-                                           sizeof(buf),
-                                           0x30,
-                                           0x01,
-                                           "New",
-                                           "Old",
-                                           true,
-                                           true,
-                                           0x00,
-                                           false,
-                                           false /* no hmac_proof */);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_ledger_account(buf, len));
-}
-
-static void test_ela_ethereum_missing_chain_id(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_edit_ledger_account(buf,
-                                           sizeof(buf),
-                                           0x30,
-                                           0x01,
-                                           "New",
-                                           "Old",
-                                           true,
-                                           true,
-                                           0x01,  /* FAMILY_ETHEREUM */
-                                           false, /* no chain_id */
-                                           true);
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_ledger_account(buf, len));
-}
-
-static void test_ela_app_rejects(void)
-{
-    g_mock_ela_result = false;
-    uint8_t buf[512];
-    size_t  len = build_valid_edit_ledger_account(buf, sizeof(buf));
-    /* edit_ledger_account returns SWO_INCORRECT_DATA (not SWO_WRONG_PARAMETER_VALUE)
-     * when the coin-app callback rejects — see ledger_account_edit.c. */
-    TEST_ASSERT_EQUAL_INT(SWO_INCORRECT_DATA, edit_ledger_account(buf, len));
-}
-
-static void test_ela_success(void)
-{
-    uint8_t buf[512];
-    size_t  len = build_valid_edit_ledger_account(buf, sizeof(buf));
-    TEST_ASSERT_EQUAL_INT(SWO_NO_RESPONSE, edit_ledger_account(buf, len));
-}
-
-static void test_ela_review_rejected(void)
-{
-    g_mock_review_choice = false;
-    uint8_t buf[512];
-    size_t  len = build_valid_edit_ledger_account(buf, sizeof(buf));
-    TEST_ASSERT_EQUAL_INT(SWO_NO_RESPONSE, edit_ledger_account(buf, len));
 }
 
 /* ── Test runner ─────────────────────────────────────────────────────────── */
@@ -1085,22 +735,6 @@ int main(void)
     RUN_TEST(test_es_hmac_fails);
     RUN_TEST(test_es_success);
     RUN_TEST(test_es_review_rejected);
-
-    /* register_ledger_account */
-    RUN_TEST(test_rla_wrong_struct_type);
-    RUN_TEST(test_rla_missing_mandatory_field);
-    RUN_TEST(test_rla_ethereum_missing_chain_id);
-    RUN_TEST(test_rla_app_rejects);
-    RUN_TEST(test_rla_success);
-    RUN_TEST(test_rla_review_rejected);
-
-    /* edit_ledger_account */
-    RUN_TEST(test_ela_wrong_struct_type);
-    RUN_TEST(test_ela_missing_mandatory_field);
-    RUN_TEST(test_ela_ethereum_missing_chain_id);
-    RUN_TEST(test_ela_app_rejects);
-    RUN_TEST(test_ela_success);
-    RUN_TEST(test_ela_review_rejected);
 
     return UNITY_END();
 }
